👌 move e2e tests related to session renew/expire in sessions.scenario
